Minimum example how SecureStore fails
When app starts, it tries to read value from SecureStore.
On initial load, it is empty, as is expected.
Once app is loaded, pressing button saves value to SecureStore.
Pressing another button loads it, and shows it is saved correctly.
Restarting app, and saved value is read correctly, and visible.
If expo-cli is restarted, then value is lost.